l10n_br_account: alterada a visão dos dados adicionais da nota fiscal de entrada e saida, resolvido métodos onchange da localização, l10n_br_delivery: criado menu para veiculos, otimizada a visão de veiculos, l10n_brs: Adicionada a pasta test dos modulos da localização para implementação de testes
